fixed off by one issue

+++ src/GenomicRegionCollection.cpp
@@ -408,7 +408,7 @@ GenomicRegionCollection<T>::GenomicRegionCollection(const T& gr)
 template<class T>
 GRC GenomicRegionCollection<T>::intersection(GRC& subject, bool ignore_strand /* false */)
 {
-  std::vector<size_t> sub, que;
+  std::vector<int32_t> sub, que;
   GRC out = this->findOverlaps(subject, que, sub, ignore_strand);
   return out;
 }
